fix: Resolve 404 error on dynamic agent pages
This commit fixes a 404 error that occurred when accessing dynamic agent pages (e.g., `/en/agent/ali-ahmadi`).

The root cause of the issue was that the agent page was implemented as a client component (`"use client"`) without a `generateStaticParams` function, preventing Next.js from pre-rendering the dynamic routes at build time.

The fix involves refactoring the agent page structure:
- The main page at `src/app/[lng]/agent/[slug]/page.tsx` is now a server component.
- It correctly implements `generateStaticParams` to generate all possible agent routes for all supported languages.
- It fetches agent data on the server side.
- A new client component, `AgentProfileCard`, has been created at `src/components/AgentClientPage.tsx` to handle client-side interactivity (like the QR code button), ensuring a proper separation of concerns.

This change ensures that all dynamic agent pages are correctly generated during the build process, resolving the 404 error and making the feature fully functional.
