+++ src/app/[lng]/agent/[slug]/page.tsx
@@ -1,106 +1,58 @@
-"use client";
-
-import { useEffect, useState } from 'react';
 import { notFound } from 'next/navigation';
 import Image from 'next/image';
-import { QRCodeCanvas as QRCode } from 'qrcode.react';
-import { Facebook, Twitter, Linkedin, Instagram, Link as LinkIcon, QrCode } from 'lucide-react';
+import { Metadata } from 'next';
 
 import Header from '@/components/Header';
 import Footer from '@/components/Footer';
-import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { useTranslation } from '@/i18n/client';
-import { agents, getAgentBySlug, Agent } from '@/lib/agents';
+import { AgentProfileCard } from '@/components/AgentClientPage'; // Importing the new client component
+import { useTranslation } from '@/i18n';
+import { languages } from '@/i18n/settings';
+import { agents, getAgentBySlug } from '@/lib/agents';
 
-// This function is commented out because it requires server-side logic in a "use client" component.
-// We will fetch the agent data on the client side instead.
-// export async function generateStaticParams() {
-//   return agents.map((agent) => ({
-//     slug: agent.slug,
-//   }));
-// }
-
-const AgentPage = ({ params: { lng, slug } }: { params: { lng: string, slug:string } }) => {
-  const { t } = useTranslation(lng, 'agent');
-  const [agent, setAgent] = useState<Agent | undefined>(undefined);
-  const [pageUrl, setPageUrl] = useState('');
-  const [showQr, setShowQr] = useState(false);
-
-  useEffect(() => {
-    const fetchedAgent = getAgentBySlug(slug);
-    setAgent(fetchedAgent);
-    if (typeof window !== 'undefined') {
-      setPageUrl(window.location.href);
+// Generate static paths for all agents and languages
+export async function generateStaticParams() {
+  const params: { lng: string; slug: string }[] = [];
+  for (const lng of languages) {
+    for (const agent of agents) {
+      params.push({ lng, slug: agent.slug });
     }
-  }, [slug]);
+  }
+  return params;
+}
+
+// Generate metadata for the page
+export async function generateMetadata({ params: { slug } }: { params: { slug: string } }): Promise<Metadata> {
+  const agent = getAgentBySlug(slug);
+  if (!agent) {
+    return { title: 'Agent Not Found' };
+  }
+  return {
+    title: `${agent.name} - Real Estate Agent`,
+    description: `Business card and portfolio for ${agent.name}, ${agent.title}.`,
+  };
+}
+
+const AgentPage = async ({ params: { lng, slug } }: { params: { lng: string, slug: string } }) => {
+  const agent = getAgentBySlug(slug);
+  const { t } = await useTranslation(lng, 'agent');
 
   if (!agent) {
-    // Since getAgentBySlug is synchronous, we can decide to show a notFound page
-    // almost immediately if the agent isn't in our static list.
-    // In a real-world scenario with data fetching, you'd have a loading state.
-    return notFound();
+    notFound();
   }
-
-  const socialIcons = {
-    Facebook: <Facebook className="h-6 w-6" />,
-    Twitter: <Twitter className="h-6 w-6" />,
-    LinkedIn: <Linkedin className="h-6 w-6" />,
-    Instagram: <Instagram className="h-6 w-6" />,
-  };
 
   return (
     <div className="min-h-screen bg-background text-foreground">
       <Header lng={lng} />
 
-      <main className="container mx-auto px-4 py-12 md:py-20">
+      <main className="container mx-auto px-4 py-28 md:py-32">
         <div className="grid grid-cols-1 md:grid-cols-3 gap-8 md:gap-12">
-          {/* Left Column: Profile Card */}
+          {/* Left Column: Profile Card (Client Component) */}
           <div className="md:col-span-1">
-            <Card className="glass sticky top-24">
-              <CardContent className="p-6 text-center">
-                <div className="relative w-40 h-40 mx-auto mb-4">
-                  <Image
-                    src={agent.imageUrl}
-                    alt={agent.name}
-                    layout="fill"
-                    objectFit="cover"
-                    className="rounded-full border-4 border-accent"
-                  />
-                </div>
-                <h1 className="text-3xl font-bold text-gradient">{agent.name}</h1>
-                <p className="text-muted-foreground text-lg mb-4">{agent.title}</p>
-
-                <div className="flex justify-center space-x-4 mb-6">
-                  {agent.socials.map((social) => (
-                    <a
-                      key={social.name}
-                      href={social.url}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="text-muted-foreground hover:text-accent transition-colors"
-                      aria-label={social.name}
-                    >
-                      {socialIcons[social.name]}
-                    </a>
-                  ))}
-                </div>
-
-                <Button onClick={() => setShowQr(!showQr)} className="w-full">
-                  <QrCode className="mr-2 h-5 w-5" />
-                  {showQr ? t('hideQr') : t('showQr')}
-                </Button>
-
-                {showQr && pageUrl && (
-                  <div className="mt-6 p-4 bg-white rounded-lg inline-block">
-                    <QRCode value={pageUrl} size={160} />
-                  </div>
-                )}
-              </CardContent>
-            </Card>
+            <AgentProfileCard agent={agent} lng={lng} />
           </div>
 
-          {/* Right Column: Portfolio */}
+          {/* Right Column: Portfolio (Server Component) */}
           <div className="md:col-span-2">
             <Card className="glass">
               <CardHeader>
@@ -111,11 +63,11 @@ const AgentPage = ({ params: { lng, slug } }: { params: { lng: string, slug:stri
                   {agent.portfolio.map((item, index) => (
                     <div key={index} className="grid grid-cols-1 sm:grid-cols-3 gap-6 items-center">
                       <div className="sm:col-span-1">
-                        <div className="aspect-w-16 aspect-h-9 relative rounded-lg overflow-hidden">
+                        <div className="relative aspect-w-16 aspect-h-9 rounded-lg overflow-hidden">
                           <Image
                             src={item.imageUrl}
                             alt={item.title}
-                            layout="fill"
+                            fill
                             objectFit="cover"
                           />
                         </div>
